requirements.txt
# Web scraping
requests==2.32.3
playwright==1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Document extraction
pdfplumber==0.11.4
pymupdf==1.24.14
python-docx==1.1.2

# AI
anthropic==0.42.0

# Database
SQLAlchemy==2.0.36

# UI
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
rich==13.9.4
